# cellphones_service/swagger_server/models/api_response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.product import Product  # noqa: F401,E501
from swagger_server import util


class ApiResponse(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, total: int=None, cached: bool=None, data: List[Product]=None):  # noqa: E501
        """ApiResponse - a model defined in Swagger

        :param total: The total of this ApiResponse.  # noqa: E501
        :type total: int
        :param cached: The cached of this ApiResponse.  # noqa: E501
        :type cached: bool
        :param data: The data of this ApiResponse.  # noqa: E501
        :type data: List[Product]
        """
        self.swagger_types = {
            'total': int,
            'cached': bool,
            'data': List[Product]
        }

        self.attribute_map = {
            'total': 'total',
            'cached': 'cached',
            'data': 'data'
        }
        self._total = total
        self._cached = cached
        self._data = data

    @classmethod
    def from_dict(cls, dikt) -> 'ApiResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ApiResponse of this ApiResponse.  # noqa: E501
        :rtype: ApiResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def total(self) -> int:
        """Gets the total of this ApiResponse.


        :return: The total of this ApiResponse.
        :rtype: int
        """
        return self._total

    @total.setter
    def total(self, total: int):
        """Sets the total of this ApiResponse.


        :param total: The total of this ApiResponse.
        :type total: int
        """

        self._total = total

    @property
    def cached(self) -> bool:
        """Gets the cached of this ApiResponse.


        :return: The cached of this ApiResponse.
        :rtype: bool
        """
        return self._cached

    @cached.setter
    def cached(self, cached: bool):
        """Sets the cached of this ApiResponse.


        :param cached: The cached of this ApiResponse.
        :type cached: bool
        """

        self._cached = cached

    @property
    def data(self) -> List[Product]:
        """Gets the data of this ApiResponse.


        :return: The data of this ApiResponse.
        :rtype: List[Product]
        """
        return self._data

    @data.setter
    def data(self, data: List[Product]):
        """Sets the data of this ApiResponse.


        :param data: The data of this ApiResponse.
        :type data: List[Product]
        """

        self._data = data